+++ tbot/main.py
@@ -1,4 +1,4 @@
-# TBot, Embedded Automation Tool
+# tbot, Embedded Automation Tool
 # Copyright (C) 2018  Harald Seiler
 #
 # This program is free software: you can redistribute it and/or modify
@@ -172,7 +172,7 @@ def main() -> None:  # noqa: C901
     if args.interactive:
         log.INTERACTIVE = True
 
-    print(log.c("TBot").yellow.bold + " starting ...")
+    print(log.c("tbot").yellow.bold + " starting ...")
 
     import tbot